AiResume2/requirements.txt
streamlit
fpdf
googletrans==4.0.0-rc1
gtts
pandas
numpy
matplotlib
seaborn
python-dotenv
PyPDF2
docx2txt
pillow
plotly
scikit-learn
spacy
transformers
torch
langchain
openai